Inline NotificationPoller into CloudMessagingHelper
It doesn't provide any additional abstraction or encapsulation.
